+++ catalog/app/containers/Admin/Buckets/Buckets.tsx
@@ -23,7 +23,6 @@ import * as GQL from 'utils/GraphQL'
 import MetaTitle from 'utils/MetaTitle'
 import * as NamedRoutes from 'utils/NamedRoutes'
 import StyledLink from 'utils/StyledLink'
-import StyledTooltip from 'utils/StyledTooltip'
 import assertNever from 'utils/assertNever'
 import parseSearch from 'utils/parseSearch'
 import { useTracker } from 'utils/tracking'
@@ -531,11 +530,11 @@ function Hint({ children }: HintProps) {
   const classes = useHintStyles()
   const tooltipClasses = React.useMemo(() => ({ tooltip: classes.tooltip }), [classes])
   return (
-    <StyledTooltip arrow maxWidth="md" title={children} classes={tooltipClasses}>
+    <M.Tooltip arrow title={children} classes={tooltipClasses}>
       <M.Icon fontSize="small" className={classes.icon}>
         help
       </M.Icon>
-    </StyledTooltip>
+    </M.Tooltip>
   )
 }
 
@@ -1716,7 +1715,7 @@ function EditPage({ back }: EditPageProps) {
       try {
         const { bucketUpdate: r } = await update({ name: bucket.name, input })
         if (r.__typename !== 'BucketUpdateSuccess') {
-          // Generated `InputError` lacks optional properties and not infered correctly
+          // Generated `InputError` lacks optional properties and not inferred correctly
           return r as Exclude<
             Model.GQLTypes.BucketUpdateResult,
             Model.GQLTypes.BucketUpdateSuccess
@@ -1756,7 +1755,7 @@ function AddPage({ back }: AddPageProps) {
       try {
         const { bucketAdd: r } = await add({ input })
         if (r.__typename !== 'BucketAddSuccess') {
-          // TS infered shape but not the actual type
+          // TS inferred shape but not the actual type
           return r as Exclude<
             Model.GQLTypes.BucketAddResult,
             Model.GQLTypes.BucketAddSuccess