+++ catalog/app/containers/UriResolver/UriResolver.tsx
@@ -1,17 +1,21 @@
 import * as React from 'react'
-import { Redirect, useHistory, useParams } from 'react-router-dom'
+import { useHistory, useParams } from 'react-router-dom'
 import * as M from '@material-ui/core'
 
 import Layout from 'components/Layout'
 import MetaTitle from 'utils/MetaTitle'
 import * as NamedRoutes from 'utils/NamedRoutes'
 import * as PackageUri from 'utils/PackageUri'
+import { BaseError } from 'utils/error'
+
+import Redirect from './Redirect'
+import parsePackageUriSafe from './parsePackageUriSafe'
 
 const useStyles = M.makeStyles((t) => ({
   container: {
-    paddingTop: t.spacing(6),
-    paddingBottom: t.spacing(6),
     maxWidth: '656px !important',
+    paddingBottom: t.spacing(2),
+    paddingTop: t.spacing(2),
   },
   form: {
     display: 'flex',
@@ -22,23 +26,17 @@ const useStyles = M.makeStyles((t) => ({
   },
 }))
 
-export default function UriResolver() {
-  const params = useParams<{ uri?: string }>()
+interface FormProps {
+  initialValue: string
+  error: PackageUri.PackageUriError | null
+}
 
+function Form({ initialValue, error }: FormProps) {
   const { urls } = NamedRoutes.use()
   const history = useHistory()
   const classes = useStyles()
 
-  const uri = decodeURIComponent(params.uri || '')
-  const [parsed, error] = React.useMemo(() => {
-    try {
-      return [uri ? PackageUri.parse(uri) : null]
-    } catch (e) {
-      return [null, e as unknown as PackageUri.PackageUriError]
-    }
-  }, [uri])
-
-  const [value, setValue] = React.useState(uri)
+  const [value, setValue] = React.useState(initialValue)
 
   const handleChange = React.useCallback(
     (e) => {
@@ -50,59 +48,64 @@ export default function UriResolver() {
   const handleSubmit = React.useCallback(
     (e) => {
       e.preventDefault()
-      if (value !== uri) history.push(urls.uriResolver(value))
+      if (value !== initialValue) history.push(urls.uriResolver(value))
     },
-    [value, uri, history, urls],
+    [value, initialValue, history, urls],
   )
 
-  const to =
-    parsed &&
-    urls.bucketPackageTree(
-      parsed.bucket,
-      parsed.name,
-      parsed.hash || parsed.tag,
-      parsed.path,
-    ) + NamedRoutes.mkSearch({ resolvedFrom: uri })
+  return (
+    <M.Container className={classes.container}>
+      <MetaTitle>Resolve a Quilt+ URI</MetaTitle>
 
-  if (to) return <Redirect to={to} />
+      <M.Typography variant="h4" align="center">
+        Resolve a Quilt+ URI
+      </M.Typography>
+
+      <form className={classes.form} onSubmit={handleSubmit}>
+        <M.Input
+          value={value}
+          onChange={handleChange}
+          error={!!error}
+          placeholder="Enter a URI, e.g. quilt+s3://your-bucket#package=user/package@hash"
+          fullWidth
+        />
+        <M.Button
+          className={classes.btn}
+          type="submit"
+          variant="contained"
+          color="primary"
+        >
+          Resolve
+        </M.Button>
+      </form>
+
+      {!!error && (
+        <M.Box mt={2}>
+          <M.Typography color="error" data-testid="uri-error">
+            Error parsing URI: {error.msg || `${error}`}
+          </M.Typography>
+        </M.Box>
+      )}
+    </M.Container>
+  )
+}
+
+export default function UriResolver() {
+  const params = useParams<{ uri?: string }>()
+
+  const decoded = decodeURIComponent(params.uri || '')
+  const uri = React.useMemo(
+    () => (decoded ? parsePackageUriSafe(decoded) : null),
+    [decoded],
+  )
+
+  if (uri && !(uri instanceof BaseError)) {
+    return <Redirect parsed={uri} decoded={decoded} />
+  }
 
   return (
-    <Layout
-      pre={
-        <M.Container className={classes.container}>
-          <MetaTitle>Resolve a Quilt+ URI</MetaTitle>
-
-          <M.Typography variant="h4" align="center">
-            Resolve a Quilt+ URI
-          </M.Typography>
-
-          <form className={classes.form} onSubmit={handleSubmit}>
-            <M.Input
-              value={value}
-              onChange={handleChange}
-              error={!!error}
-              placeholder="Enter a URI, e.g. quilt+s3://your-bucket#package=user/package@hash"
-              fullWidth
-            />
-            <M.Button
-              className={classes.btn}
-              type="submit"
-              variant="contained"
-              color="primary"
-            >
-              Resolve
-            </M.Button>
-          </form>
-
-          {!!error && (
-            <M.Box mt={2}>
-              <M.Typography color="error" data-testid="uri-error">
-                Error parsing URI: {error.msg || `${error}`}
-              </M.Typography>
-            </M.Box>
-          )}
-        </M.Container>
-      }
-    />
+    <Layout>
+      <Form initialValue={decoded} error={uri} />
+    </Layout>
   )
 }